refactor: Controller 생성자의 주입 객체 변경
