+++ src/main/java/bridge/Application.java
@@ -1,13 +1,12 @@
 package bridge;
 
 import bridge.controller.BridgeController;
-import bridge.view.InputView;
-import bridge.view.OutputView;
+import bridge.domain.generator.BridgeRandomNumberGenerator;
 
 public class Application {
 
     public static void main(String[] args) {
-        BridgeController bridgeController = new BridgeController();
+        BridgeController bridgeController = new BridgeController(new BridgeRandomNumberGenerator());
         bridgeController.start();
     }
 }
