+++ src/main/java/bridge/domain/generator/BridgeRandomNumberGenerator.java
@@ -1,6 +1,6 @@
-package bridge.domain;
+package bridge.domain.generator;
 
-import bridge.domain.BridgeNumberGenerator;
+import bridge.domain.generator.BridgeNumberGenerator;
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class BridgeRandomNumberGenerator implements BridgeNumberGenerator {
